Fix clippy lints for rustc 1.94.1 and remove unnecessary #[allow] attributes
Replace needless_range_loop with iter/iter_mut + enumerate patterns,
fix doc overindent, remove dead_code allows on actively-used structs.

+++ crates/celox/src/backend/native/regalloc/tests.rs
@@ -137,7 +137,7 @@ fn test_high_pressure_spill() {
     // Use all 14 in a chain of adds
     let mut acc = VReg(0);
     for i in 1..14 {
-        let dst = VReg(14 + i as u32);
+        let dst = VReg(14 + i);
         insts.push(MInst::Add {
             dst,
             lhs: acc,
